+++ feature/src/main/java/com/teamkkumul/feature/meetup/MeetUpDetailVpAdapter.kt
@@ -1,10 +1,10 @@
-package com.teamkkumul.feature.home.meetupdetail
+package com.teamkkumul.feature.meetup
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.teamkkumul.feature.home.meetupdetail.lateperson.LatePersonFragment
-import com.teamkkumul.feature.home.meetupdetail.readystatus.ReadyStatusFragment
-import com.teamkkumul.feature.home.meetupdetail.meetupinfo.MeetUpDetailFragment
+import com.teamkkumul.feature.meetup.lateperson.LatePersonFragment
+import com.teamkkumul.feature.meetup.meetupdetail.MeetUpDetailFragment
+import com.teamkkumul.feature.meetup.readystatus.ReadyStatusFragment
 
 class MeetUpDetailVpAdapter(fr: Fragment) : FragmentStateAdapter(fr) {
     override fun getItemCount(): Int = TOTAL_TAG_NUM
