+++ feature/src/main/java/com/teamkkumul/feature/meetup/meetupdetail/MeetUpDetailFriendViewModel.kt
@@ -1,4 +1,4 @@
-package com.teamkkumul.feature.home.meetupdetail.meetupinfo
+package com.teamkkumul.feature.meetup.meetupdetail
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
