+++ feature/src/main/java/com/teamkkumul/feature/meetup/meetupdetail/viewholder/MeetUpDetailFriendPlusViewHolder.kt
@@ -1,4 +1,4 @@
-package com.teamkkumul.feature.home.meetupdetail.meetupinfo.viewholder
+package com.teamkkumul.feature.meetup.meetupdetail.viewholder
 
 import androidx.recyclerview.widget.RecyclerView
 import com.teamkkumul.feature.databinding.ItemMyGroupFriendPlusBinding
@@ -7,8 +7,5 @@ import com.teamkkumul.model.MeetUpSealedItem
 class MeetUpDetailFriendPlusViewHolder(
     private val binding: ItemMyGroupFriendPlusBinding,
 ) : RecyclerView.ViewHolder(binding.root) {
-    init {
-    }
-
     fun onBind(data: MeetUpSealedItem.MyGroupPlus) {}
 }
