+++ feature/src/main/res/navigation/nav_graph.xml
@@ -45,7 +45,7 @@
 
     <fragment
         android:id="@+id/fragment_meet_up_detail"
-        android:name="com.teamkkumul.feature.home.meetupdetail.meetupinfo.MeetUpDetailFragment"
+        android:name="com.teamkkumul.feature.meetup.meetupdetail.MeetUpDetailFragment"
         android:label="My Group Detail"
         tools:layout="@layout/fragment_meet_up_detail" />
 
@@ -82,7 +82,7 @@
         android:label="WelcomeFragment" />
     <fragment
         android:id="@+id/meetUpContainerFragment"
-        android:name="com.teamkkumul.feature.home.meetupdetail.MeetUpContainerFragment"
+        android:name="com.teamkkumul.feature.meetup.MeetUpContainerFragment"
         android:label="MeetUpContainerFragment" />
 
 </navigation>
